+++ src/OpenTelemetry/Metrics/MetricPoint.cs
@@ -15,18 +15,20 @@
 // </copyright>
 
 using System;
+using System.Diagnostics;
 using System.Threading;
 
 namespace OpenTelemetry.Metrics
 {
     public struct MetricPoint
     {
+        private readonly AggregationType aggType;
+        private readonly object lockObject;
+        private readonly long[] bucketCounts;
         private long longVal;
         private long lastLongSum;
         private double doubleVal;
         private double lastDoubleSum;
-        private object lockObject;
-        private long[] bucketCounts;
 
         internal MetricPoint(
             AggregationType aggType,
@@ -35,10 +37,11 @@ namespace OpenTelemetry.Metrics
             object[] values,
             double[] histogramBounds)
         {
-            this.AggType = aggType;
+            Debug.Assert((keys?.Length ?? 0) == (values?.Length ?? 0), "Key and value array lengths did not match.");
+
+            this.aggType = aggType;
             this.StartTime = startTime;
-            this.Keys = keys;
-            this.Values = values;
+            this.Tags = new ReadOnlyTagCollection(keys, values);
             this.EndTime = default;
             this.LongValue = default;
             this.longVal = default;
@@ -48,14 +51,14 @@ namespace OpenTelemetry.Metrics
             this.lastDoubleSum = default;
             this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
-            if (this.AggType == AggregationType.Histogram)
+            if (this.aggType == AggregationType.Histogram)
             {
                 this.ExplicitBounds = histogramBounds;
                 this.BucketCounts = new long[this.ExplicitBounds.Length + 1];
                 this.bucketCounts = new long[this.ExplicitBounds.Length + 1];
                 this.lockObject = new object();
             }
-            else if (this.AggType == AggregationType.HistogramSumCount)
+            else if (this.aggType == AggregationType.HistogramSumCount)
             {
                 this.ExplicitBounds = null;
                 this.BucketCounts = null;
@@ -71,9 +74,10 @@ namespace OpenTelemetry.Metrics
             }
         }
 
-        public string[] Keys { get; internal set; }
-
-        public object[] Values { get; internal set; }
+        /// <summary>
+        /// Gets the tags associated with the metric point.
+        /// </summary>
+        public ReadOnlyTagCollection Tags { get; }
 
         public DateTimeOffset StartTime { get; internal set; }
 
@@ -89,11 +93,9 @@ namespace OpenTelemetry.Metrics
 
         internal MetricPointStatus MetricPointStatus { get; private set; }
 
-        private readonly AggregationType AggType { get; }
-
         internal void Update(long number)
         {
-            switch (this.AggType)
+            switch (this.aggType)
             {
                 case AggregationType.LongSumIncomingDelta:
                     {
@@ -137,7 +139,7 @@ namespace OpenTelemetry.Metrics
 
         internal void Update(double number)
         {
-            switch (this.AggType)
+            switch (this.aggType)
             {
                 case AggregationType.DoubleSumIncomingDelta:
                     {
@@ -213,7 +215,7 @@ namespace OpenTelemetry.Metrics
 
         internal void TakeSnapshot(bool outputDelta)
         {
-            switch (this.AggType)
+            switch (this.aggType)
             {
                 case AggregationType.LongSumIncomingDelta:
                 case AggregationType.LongSumIncomingCumulative: