+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Added `ReadOnlyTagCollection` and expose `Tags` on `MetricPoint` instead of
+  `Keys`+`Values`
+  ([#2642](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2642))
+
 ## 1.2.0-beta2
 
 Released 2021-Nov-19